污水处理

+++ internal/app/handlers/request/pasture.go
@@ -21,8 +21,8 @@ type ReqEndFeeding struct {
 	BatchNumber string  `json:"batch_number" form:"batch_number"`
 	Worker      string  `json:"worker" form:"worker"`
 	HouseNumber string  `json:"house_number" form:"house_number"`
-	PM10        float32 `json:"pm_10" form:"pm_10"`   // PM10
-	TSP         float32 `json:"tsp" form:"tsp"`       // TSP
+	PM10        float64 `json:"pm_10" form:"pm_10"`   // PM10
+	TSP         float64 `json:"tsp" form:"tsp"`       // TSP
 	Stench      uint    `json:"stench" form:"stench"` // 恶臭稀释程度
 }
 
@@ -62,90 +62,90 @@ type ReqAddPastureWaterRecord struct {
 type ReqAddPastureBuffer struct {
 	HouseNumber string  `json:"houseNumber"`
 	TimeStamp   int64   `json:"time_stamp"`
-	Buffer1     float32 `json:"buffer_1"`  //NH3氨气
-	Buffer2     float32 `json:"buffer_2"`  //H2S硫化氢
-	Buffer3     float32 `json:"buffer_3"`  //CO2二氧化碳
-	Buffer4     float32 `json:"buffer_4"`  //PM10
-	Buffer5     float32 `json:"buffer_5"`  //TSP
-	Buffer6     float32 `json:"buffer_6"`  //ODO 恶臭
-	Buffer7     float32 `json:"buffer_7"`  //Cd(镉)
-	Buffer8     float32 `json:"buffer_8"`  //As（砷）
-	Buffer9     float32 `json:"buffer_9"`  //Cu（铜）
-	Buffer10    float32 `json:"buffer_10"` //Pb（铅）
-	Buffer11    float32 `json:"buffer_11"` //Cr(铬)
-	Buffer12    float32 `json:"buffer_12"` //Zn(锌）
+	Buffer1     float64 `json:"buffer_1"`  //NH3氨气
+	Buffer2     float64 `json:"buffer_2"`  //H2S硫化氢
+	Buffer3     float64 `json:"buffer_3"`  //CO2二氧化碳
+	Buffer4     float64 `json:"buffer_4"`  //PM10
+	Buffer5     float64 `json:"buffer_5"`  //TSP
+	Buffer6     float64 `json:"buffer_6"`  //ODO 恶臭
+	Buffer7     float64 `json:"buffer_7"`  //Cd(镉)
+	Buffer8     float64 `json:"buffer_8"`  //As（砷）
+	Buffer9     float64 `json:"buffer_9"`  //Cu（铜）
+	Buffer10    float64 `json:"buffer_10"` //Pb（铅）
+	Buffer11    float64 `json:"buffer_11"` //Cr(铬)
+	Buffer12    float64 `json:"buffer_12"` //Zn(锌）
 }
 
 type ReqAddPastureArea struct {
 	TimeStamp    int64   `json:"time_stamp"`
 	HouseNumber  string  `gorm:"not null; type:varchar(256)" json:"house_number"`
-	CattleFarm1  float32 `json:"cattle_farm_1"`  //NH3氨气
-	CattleFarm2  float32 `json:"cattle_farm_2"`  //H2S硫化氢
-	CattleFarm3  float32 `json:"cattle_farm_3"`  //CO2二氧化碳
-	CattleFarm4  float32 `json:"cattle_farm_4"`  //PM10
-	CattleFarm5  float32 `json:"cattle_farm_5"`  //TSP
-	CattleFarm6  float32 `json:"cattle_farm_6"`  //ODO 恶臭
-	CattleFarm7  float32 `json:"cattle_farm_7"`  //Cd(镉)
-	CattleFarm8  float32 `json:"cattle_farm_8"`  //As（砷）
-	CattleFarm9  float32 `json:"cattle_farm_9"`  //Cu（铜）
-	CattleFarm10 float32 `json:"cattle_farm_10"` //Pb（铅）
-	CattleFarm11 float32 `json:"cattle_farm_11"` //Cr(铬)
-	CattleFarm12 float32 `json:"cattle_farm_12"` //Zn(锌）
+	CattleFarm1  float64 `json:"cattle_farm_1"`  //NH3氨气
+	CattleFarm2  float64 `json:"cattle_farm_2"`  //H2S硫化氢
+	CattleFarm3  float64 `json:"cattle_farm_3"`  //CO2二氧化碳
+	CattleFarm4  float64 `json:"cattle_farm_4"`  //PM10
+	CattleFarm5  float64 `json:"cattle_farm_5"`  //TSP
+	CattleFarm6  float64 `json:"cattle_farm_6"`  //ODO 恶臭
+	CattleFarm7  float64 `json:"cattle_farm_7"`  //Cd(镉)
+	CattleFarm8  float64 `json:"cattle_farm_8"`  //As（砷）
+	CattleFarm9  float64 `json:"cattle_farm_9"`  //Cu（铜）
+	CattleFarm10 float64 `json:"cattle_farm_10"` //Pb（铅）
+	CattleFarm11 float64 `json:"cattle_farm_11"` //Cr(铬)
+	CattleFarm12 float64 `json:"cattle_farm_12"` //Zn(锌）
 }
 
 type ReqAddCowHouse struct {
 	TimeStamp   int64   `json:"time_stamp"`
 	HouseNumber string  `gorm:"not null; type:varchar(256)" json:"house_number"`
-	CowHouse1   float32 `json:"cow_house_1"`  //NH3氨气
-	CowHouse2   float32 `json:"cow_house_2"`  //H2S硫化氢
-	CowHouse3   float32 `json:"cow_house_3"`  //CO2二氧化碳
-	CowHouse4   float32 `json:"cow_house_4"`  //PM10
-	CowHouse5   float32 `json:"cow_house_5"`  //TSP
-	CowHouse6   float32 `json:"cow_house_6"`  //ODO 恶臭
-	CowHouse7   float32 `json:"cow_house_7"`  //Cd(镉)
-	CowHouse8   float32 `json:"cow_house_8"`  //As（砷）
-	CowHouse9   float32 `json:"cow_house_9"`  //Cu（铜）
-	CowHouse10  float32 `json:"cow_house_10"` //Pb（铅）
-	CowHouse11  float32 `json:"cow_house_11"` //Cr(铬)
-	CowHouse12  float32 `json:"cow_house_12"` //Zn(锌）
+	CowHouse1   float64 `json:"cow_house_1"`  //NH3氨气
+	CowHouse2   float64 `json:"cow_house_2"`  //H2S硫化氢
+	CowHouse3   float64 `json:"cow_house_3"`  //CO2二氧化碳
+	CowHouse4   float64 `json:"cow_house_4"`  //PM10
+	CowHouse5   float64 `json:"cow_house_5"`  //TSP
+	CowHouse6   float64 `json:"cow_house_6"`  //ODO 恶臭
+	CowHouse7   float64 `json:"cow_house_7"`  //Cd(镉)
+	CowHouse8   float64 `json:"cow_house_8"`  //As（砷）
+	CowHouse9   float64 `json:"cow_house_9"`  //Cu（铜）
+	CowHouse10  float64 `json:"cow_house_10"` //Pb（铅）
+	CowHouse11  float64 `json:"cow_house_11"` //Cr(铬)
+	CowHouse12  float64 `json:"cow_house_12"` //Zn(锌）
 }
 
 type ReqAddPastureBasicEnvironment struct {
 	TimeStamp    int64   `json:"time_stamp"`
 	HouseNumber  string  `gorm:"not null; type:varchar(256)" json:"house_number"`
-	Environment1 float32 `json:"environment_1"` //温度
-	Environment2 float32 `json:"environment_2"` //相对湿度
-	Environment3 float32 `json:"environment_3"` //风速
-	Environment4 float32 `json:"environment_4"` //照度
-	Environment5 float32 `json:"environment_5"` //噪声
-	Environment6 float32 `json:"environment_6"` //光照时间
+	Environment1 float64 `json:"environment_1"` //温度
+	Environment2 float64 `json:"environment_2"` //相对湿度
+	Environment3 float64 `json:"environment_3"` //风速
+	Environment4 float64 `json:"environment_4"` //照度
+	Environment5 float64 `json:"environment_5"` //噪声
+	Environment6 float64 `json:"environment_6"` //光照时间
 }
 
 type ReqAddPasturePaddingRequire struct {
 	TimeStamp       int64   `json:"time_stamp"`
 	HouseNumber     string  `gorm:"not null; type:varchar(256)" json:"house_number"`
-	PaddingRequire1 float32 `json:"padding_require_1"` //Hg（总汞）
-	PaddingRequire2 float32 `json:"padding_require_2"` //Pb（铅）
-	PaddingRequire3 float32 `json:"padding_require_3"` //Cr（铬）
-	PaddingRequire4 float32 `json:"padding_require_4"` //Cd（镉）
-	PaddingRequire5 float32 `json:"padding_require_5"` //TTC(总大肠菌群数)
-	PaddingRequire6 float32 `json:"padding_require_6"` //TBC(细菌总数)
-	PaddingRequire7 float32 `json:"padding_require_7"` //AFB1（黄曲霉素B1）
-	PaddingRequire8 float32 `json:"padding_require_8"` //STC(沙门菌数)
+	PaddingRequire1 float64 `json:"padding_require_1"` //Hg（总汞）
+	PaddingRequire2 float64 `json:"padding_require_2"` //Pb（铅）
+	PaddingRequire3 float64 `json:"padding_require_3"` //Cr（铬）
+	PaddingRequire4 float64 `json:"padding_require_4"` //Cd（镉）
+	PaddingRequire5 float64 `json:"padding_require_5"` //TTC(总大肠菌群数)
+	PaddingRequire6 float64 `json:"padding_require_6"` //TBC(细菌总数)
+	PaddingRequire7 float64 `json:"padding_require_7"` //AFB1（黄曲霉素B1）
+	PaddingRequire8 float64 `json:"padding_require_8"` //STC(沙门菌数)
 }
 
 type ReqAddPastureWastedWaterIndex struct {
 	TimeStamp         int64   `json:"time_stamp"`
 	HouseNumber       string  `gorm:"not null; type:varchar(256)" json:"house_number"`
-	WastedWaterIndex1 float32 `json:"wasted_water_index_1"` //BOD(五日生化需氧量)
-	WastedWaterIndex2 float32 `json:"wasted_water_index_2"` //COD(化学需氧量)
-	WastedWaterIndex3 float32 `json:"wasted_water_index_3"` //NH3-N(氨氮)
-	WastedWaterIndex4 float32 `json:"wasted_water_index_4"` //TP(总磷)
-	WastedWaterIndex5 float32 `json:"wasted_water_index_5"` //TSS(悬浮物）
-	WastedWaterIndex6 float32 `json:"wasted_water_index_6"` //FCC（粪大肠菌群个数）
-	WastedWaterIndex7 float32 `json:"wasted_water_index_7"` //蛔虫卵个数
-	WastedWaterIndex8 float32 `json:"wasted_water_index_8"` //PH值
-	WastedWaterIndex9 float32 `json:"wasted_water_index_9"` //流量
+	WastedWaterIndex1 float64 `json:"wasted_water_index_1"` //BOD(五日生化需氧量)
+	WastedWaterIndex2 float64 `json:"wasted_water_index_2"` //COD(化学需氧量)
+	WastedWaterIndex3 float64 `json:"wasted_water_index_3"` //NH3-N(氨氮)
+	WastedWaterIndex4 float64 `json:"wasted_water_index_4"` //TP(总磷)
+	WastedWaterIndex5 float64 `json:"wasted_water_index_5"` //TSS(悬浮物）
+	WastedWaterIndex6 float64 `json:"wasted_water_index_6"` //FCC（粪大肠菌群个数）
+	WastedWaterIndex7 float64 `json:"wasted_water_index_7"` //蛔虫卵个数
+	WastedWaterIndex8 float64 `json:"wasted_water_index_8"` //PH值
+	WastedWaterIndex9 float64 `json:"wasted_water_index_9"` //流量
 }
 
 type ReqAddPastureDisinfectionRecord struct {
@@ -165,28 +165,28 @@ type ReqPastureSensorData struct {
 type ReqPastureWasteWaterPerDay struct {
 	TimeStamp                   int64   `json:"time_stamp"`
 	HouseNumber                 string  `json:"house_number"`
-	ReqPastureWasteWaterPerDay1 float32 `json:"req_pasture_waste_water_per_day_1"`
-	ReqPastureWasteWaterPerDay2 float32 `json:"req_pasture_waste_water_per_day_2"`
-	ReqPastureWasteWaterPerDay3 float32 `json:"req_pasture_waste_water_per_day_3"`
-	ReqPastureWasteWaterPerDay4 float32 `json:"req_pasture_waste_water_per_day_4"`
+	ReqPastureWasteWaterPerDay1 float64 `json:"req_pasture_waste_water_per_day_1"`
+	ReqPastureWasteWaterPerDay2 float64 `json:"req_pasture_waste_water_per_day_2"`
+	ReqPastureWasteWaterPerDay3 float64 `json:"req_pasture_waste_water_per_day_3"`
+	ReqPastureWasteWaterPerDay4 float64 `json:"req_pasture_waste_water_per_day_4"`
 }
 
 type ReqPastureWasteResiduePerDay struct {
 	TimeStamp                  int64   `json:"time_stamp"`
 	HouseNumber                string  `gorm:"not null; type:varchar(256)" json:"house_number"`
-	PastureWasteResiduePerDay1 float32 `json:"pasture_waste_residue_per_day_1"`
-	PastureWasteResiduePerDay2 float32 `json:"pasture_waste_residue_per_day_2"`
-	PastureWasteResiduePerDay3 float32 `json:"pasture_waste_residue_per_day_3"`
-	PastureWasteResiduePerDay4 float32 `json:"pasture_waste_residue_per_day_4"`
+	PastureWasteResiduePerDay1 float64 `json:"pasture_waste_residue_per_day_1"`
+	PastureWasteResiduePerDay2 float64 `json:"pasture_waste_residue_per_day_2"`
+	PastureWasteResiduePerDay3 float64 `json:"pasture_waste_residue_per_day_3"`
+	PastureWasteResiduePerDay4 float64 `json:"pasture_waste_residue_per_day_4"`
 }
 
 type ReqPastureOdorPollutantsPerDay struct {
 	TimeStamp                    int64   `json:"time_stamp"`
 	HouseNumber                  string  `gorm:"not null; type:varchar(256)" json:"house_number"`
-	PastureOdorPollutantsPerDay1 float32 `json:"pasture_odor_pollutants_per_day_1"`
-	PastureOdorPollutantsPerDay2 float32 `json:"pasture_odor_pollutants_per_day_2"`
-	PastureOdorPollutantsPerDay3 float32 `json:"pasture_odor_pollutants_per_day_3"`
-	PastureOdorPollutantsPerDay4 float32 `json:"pasture_odor_pollutants_per_day_4"`
+	PastureOdorPollutantsPerDay1 float64 `json:"pasture_odor_pollutants_per_day_1"`
+	PastureOdorPollutantsPerDay2 float64 `json:"pasture_odor_pollutants_per_day_2"`
+	PastureOdorPollutantsPerDay3 float64 `json:"pasture_odor_pollutants_per_day_3"`
+	PastureOdorPollutantsPerDay4 float64 `json:"pasture_odor_pollutants_per_day_4"`
 }
 
 type ReqWasteResidueOdor struct {
