+++ internal/app/models/pasture/floor_bedding_physical_hazard.go
@@ -4,17 +4,17 @@ import "gorm.io/gorm"
 
 type PastureFloorBeddingPhysicalHazard struct {
 	gorm.Model
-	Mercury               float32 `json:"mercury"`  // 汞
-	Cadmium               float32 `json:"cadmium"`  // 镉
-	Lead                  float32 `json:"lead"`     // 铅
-	Chromium              float32 `json:"chromium"` // 铬
+	Mercury               float64 `json:"mercury"`  // 汞
+	Cadmium               float64 `json:"cadmium"`  // 镉
+	Lead                  float64 `json:"lead"`     // 铅
+	Chromium              float64 `json:"chromium"` // 铬
 	PastureFloorBeddingID uint    `json:"pasture_floor_bedding_id"`
 }
 
 type PastureFloorBeddingPhysicalHazardData struct {
-	Mercury               float32 `json:"mercury"`  // 汞
-	Cadmium               float32 `json:"cadmium"`  // 镉
-	Lead                  float32 `json:"lead"`     // 铅
-	Chromium              float32 `json:"chromium"` // 铬
+	Mercury               float64 `json:"mercury"`  // 汞
+	Cadmium               float64 `json:"cadmium"`  // 镉
+	Lead                  float64 `json:"lead"`     // 铅
+	Chromium              float64 `json:"chromium"` // 铬
 	PastureFloorBeddingID uint    `json:"pasture_floor_bedding_id"`
 }
